requirements.txt
fastapi==0.115.12
kink==0.8.1
pendulum==3.1.0
pydantic==2.11.4
PyYAML==6.0.2
uvicorn==0.34.2
pytest==8.3.5
